pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "nest-asyncio>=1.6.0",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot==13.14",
    "sqlalchemy>=2.0.40",
    "telegram>=0.0.1",
    "uvicorn>=0.34.1",
    "werkzeug>=3.1.3",
]
